ファイル配置整理、不要なファイル削除
