+++ app/.vscode/extensions.json
@@ -1,0 +1,8 @@
+{
+  "recommendations": [
+    "ms-python.python", // python
+    "esbenp.prettier-vscode", // prettier
+    "charliermarsh.ruff", // ruff
+    "be5invis.toml" // toml
+  ]
+}
