--- app/src/cruds/__init__.py
@@ -1,3 +1,0 @@
-from src.cruds import crud_tasks
-
-__all__ = ["crud_tasks"]
