+++ app/src/main.py
@@ -1,26 +1,22 @@
 import logging
-import sys
 from typing import Any
 
 import uvicorn
 from debug_toolbar.middleware import DebugToolbarMiddleware
 from fastapi import FastAPI, Request
 from fastapi.exceptions import HTTPException, RequestValidationError
-from fastapi.responses import JSONResponse
 
 from src.api.endpoints import tasks
-from src.config import settings
-from src.core.logger import logger
+from src.core.config import settings
+from src.core.logger import setup_logging
 from src.extensions.exception_handlers import (
   custom_http_exception_handler,
   custom_validation_exception_handler,
 )
 
-# from src.logger import setup_logging
+setup_logging()
 
-# setup_logging()
-
-# logger = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
 app = FastAPI(debug=True, root_path=settings.app_root_path)
 
@@ -56,8 +52,7 @@ app.include_router(tasks.router, tags=["Tasks"], prefix="/tasks")
 
 @app.get("/")
 async def pong() -> dict[str, Any]:
-  await logger.awarning("Pinging the server")
-
+  await logger.warning("Pinging the server")
 
   try:
     raise ValueError("test error")
@@ -66,8 +61,6 @@ async def pong() -> dict[str, Any]:
 
   return {
     "msg": "I am task app api!",
-    "environment": "dev",
-    "testing": True,
   }
