+++ app/src/models/__init__.py
@@ -2,5 +2,4 @@ from .base import Base
 from .task import Task
 from .user import User
 
-
 __all__ = ["Base", "Task", "User"]
