+++ app/src/models/user.py
@@ -1,9 +1,8 @@
-from sqlalchemy import Boolean, Text
+from sqlalchemy import Boolean
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from src.models.base import Base, CommonColumns, int_pk, str_255 
+from src.models.base import Base, CommonColumns, int_pk, str_255
 
-from typing import List
 
 # Userテーブル
 class User(Base, CommonColumns):
@@ -17,8 +16,8 @@ class User(Base, CommonColumns):
   hashed_password: Mapped[str_255] = mapped_column(comment="ハッシュ化パスワード")
 
   tasks = relationship(
-        "Task", back_populates="create_user", cascade="all, delete-orphan"
-     )
-  
+    "Task", back_populates="create_user", cascade="all, delete-orphan"
+  )
+
   def __repr__(self) -> str:
     return f"User(id={self.id!r}, username={self.username!r})"
