--- index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
-  <link rel="icon" sizes="16x16" href="https://d1qyw1y8td7tdl.cloudfront.net/favicon.ico" />
-</head>
-
-<body>
-  <h1>favicon test</h1>
-</body>
-
-</html>
